fix(profile): return precise completeness percentages

+++ src/main/java/com/fpt/ibom/profile/dto/ProfileCompletenessResponse.java
@@ -1,7 +1,8 @@
 package com.fpt.ibom.profile.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record ProfileCompletenessResponse(int percentage, boolean completed,
+public record ProfileCompletenessResponse(BigDecimal percentage, boolean completed,
 		List<ProfileCompletenessSectionResponse> sections) {
 }
